pyproject.toml
[tool.poetry]
name = "learning_loop_node"
version = "0.1.0"
description = "Python Library for Nodes which connect to the Zauberzeug Learning Loop"
authors = ["Zauberzeug GmbH <info@zauberzeug.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/zauberzeug/learning_loop_node"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
python-socketio = "^5.7.2"
fastapi-socketio = "^0.0.6"
fastapi = "^0.70.1"
uvicorn = "^0.13.3"
fastapi-utils = "^0.2.1"
simplejson = "^3.17.2"
icecream = "^2.1.0"
aiofiles = "^0.7.0"
async_generator = "^1.10"
werkzeug = "^2.0.1"
pytest-watch = "^4.2.0"
python-multipart = "^0.0.5"
psutil = "^5.8.0"
numpy = "^1.13.3"
Pillow = "^10.0.0"
tqdm = "^4.63.0"
pytest-mock = "3.6.1"
pynvml = "^11.4.1"
aiohttp = "^3.8.4"
httpx = "^0.24.1"
dacite = "^1.8.1"

[tool.poetry.group.dev.dependencies]
pytest-flakefinder = "^1.1.0"
retry = "^0.9.2"
debugpy = "^1.6.7.post1"
pytest-asyncio = "^0.21.1"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
